Ubuntu 22.04 cursors
It's out now and we don't want to be looking outdated. I use MATE but it's much the same. White cursors aren't included in it at all. I dislike a lot of the changes, such as rounded corners all over the place. But I guess I'll be replacing my installation with it sooner or later, as sticking with an old OS is just asking for trouble, even if still in support.

Images are based what Chrome displays in 22.04, which in turn aligns with libGDX/LWJGL3/GLFW in 20.04 (haven't had time to test gdx-tests because of Teams keeps pinging). Firefox is still a little funky, displaying four connected arrows for `move`, but that's alright and would be more like what Windows users see anyway.