Update maven-integration.md (#177)
* Update maven-integration.md

The article contains various outdated information, alongside not displaying when the article was updated last.
Additionally, the archetype doesn't work anymore, and Maven seems to work pretty well with libgdx according to my own findings. I'll have to return when I've tested platform compatibility and update the rest.

* Version 2

Removed "last updated" as syggested by Berstanio, and instead went with making it clear what information was updated, and what wasn't.
Re-introduced the warning as suggested by crykn, but adding the exact known issues Maven runs into.
Decided to keep the exact Maven-specific dependency information as a courtesy, as as far as I'm aware, since everyone else uses Gradle, this is not described or showcased anywhere.