Set Up a Dev Env: Newer JDK and whatnot (#206)
* Capital J for Java

* Pretend non-LTS Java doesn't exist

* Remove references to AIDE as it is no longer on Play Store

I am interested in having some form of support for developing libGDX applications on a mobile phone, as developers without access to a computer, such as for financial reasons, do exist. But I haven't got anything working.

* Update Android SDK link

* AIDE re-added with a caveat

This is a bit of a "citation needed" situation, so https://discourse.android-ide.com/t/aide-remove-from-play-store/3817/10 and my inability to get it working on Android 11 back when it was on the Play Store can be my citation.